third_party/blink/web_tests/virtual/select-parser-relaxation/external/wpt/html/semantics/forms/the-select-element/customizable-select/select-picker-hover-active-pseudo.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] defaultbutton: select should match :hover and :active when interacting with button.
  assert_equals: appearance:base-select must be supported in order to run this test. expected "base-select" but got "auto"
[FAIL] defaultbutton: select should not match :hover or :active when interacting with elements in the picker.
  assert_equals: appearance:base-select must be supported in order to run this test. expected "base-select" but got "auto"
[FAIL] custombutton: select should match :hover and :active when interacting with button.
  assert_equals: appearance:base-select must be supported in order to run this test. expected "base-select" but got "auto"
[FAIL] custombutton: select should not match :hover or :active when interacting with elements in the picker.
  assert_equals: appearance:base-select must be supported in order to run this test. expected "base-select" but got "auto"
Harness: the test ran to completion.
